requirements.txt
Django==2.0.5
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.4
sqlparse==0.2.4
